@@ -20,21 +20,21 @@
 		"@babel/preset-react": "^7.14.5",
 		"@harmelodic/react-ui-lib": "^3.4.3",
 		"babel-loader": "^8.2.2",
-		"compression-webpack-plugin": "^8.0.1",
-		"eslint": "^7.32.0",
+		"compression-webpack-plugin": "^9.2.0",
+		"eslint": "^8.7.0",
 		"eslint-config-google": "^0.14.0",
 		"eslint-plugin-react": "^7.24.0",
-		"marked": "^2.1.3",
+		"marked": "^4.0.10",
 		"react": "^17.0.1",
 		"react-dom": "^17.0.1",
 		"react-redux": "^7.2.2",
-		"react-router-dom": "^5.3.0",
+		"react-router-dom": "^6.2.1",
 		"redux": "^4.0.5",
 		"redux-thunk": "^2.3.0",
 		"styled-components": "^5.2.1",
 		"webpack": "^5.39.1",
 		"webpack-cli": "^4.7.2",
-		"webpack-dev-server": "^3.11.3",
+		"webpack-dev-server": "^4.7.3",
 		"webpack-merge": "^5.8.0"
 	},
 	"repository": {
