@@ -4,10 +4,11 @@ const PUBLIC = __dirname + '/public/';
 module.exports = {
 	devServer: {
 		compress: true,
-		contentBase: PUBLIC,
+		static: {
+			directory: PUBLIC
+		},
 		historyApiFallback: true,
 		hot: true,
-		inline: true,
 	},
 	entry: SRC + 'index.js',
 	module: {
